feat(frontend): remove demo login and add landing navigation
