+++ frontend/src/pages/LoginPage.tsx
@@ -1,120 +1,81 @@
-import react,{ useState } from 'react';
-import { useForm } from 'react-hook-form';
-import { z } from 'zod';
-import { zodResolver } from '@hookform/resolvers/zod';
+import React, { useState } from 'react';
+import { useTranslation } from 'react-i18next';
+import { Link, useNavigate, useSearchParams } from 'react-router-dom';
+import { useLanguage } from '@/context/LanguageContext';
+import { useAuth } from '@/features/auth/hooks';
+import { useToast } from '@/hooks/use-toast';
+import {
+  Card, CardContent, CardHeader, CardTitle, CardDescription
+} from '@/components/ui/card';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
-import {
-  Form,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormControl,
-  FormMessage,
-} from '@/components/ui/form';
-import { useAuth } from '@/features/auth/hooks';
-import { useAuthRedirect } from '@/features/auth/useAuthRedirect';
-
-const schema = z.object({
-  email: z.string().email(),
-  password: z.string().min(6),
-});
+import { Label } from '@/components/ui/label';
+import { Checkbox } from '@/components/ui/checkbox';
+import { LanguageToggle } from '@/components/ui/language-toggle';
+import { Eye, EyeOff, ArrowRight, AlertCircle } from 'lucide-react';
+import authBackground from '@/assets/auth-background.jpg';
 
 const LoginPage: React.FC = () => {
+  const { t } = useTranslation();
+  const { isRTL } = useLanguage();
   const { login, isLoading } = useAuth();
-  const redirect = useAuthRedirect();
-  const form = useForm<z.infer<typeof schema>>({
-    resolver: zodResolver(schema),
-    defaultValues: { email: '', password: '' },
-  });
+  const navigate = useNavigate();
+  const [searchParams] = useSearchParams();
+  const { toast } = useToast();
 
-  const onSubmit = async (values: z.infer<typeof schema>) => {
-    const success = await login(values);
-    if (success) {
-      redirect();
-    }
-  };
-
- 
- 
   const [formData, setFormData] = useState({
     email: '',
     password: '',
-    remember: false
-  }); 
-  const validateForm = () => {
-    const newErrors: Record<string, string> = {};
-
-    if (!formData.email) {
-      newErrors.email = t('auth.validation.required');
-    } else if (!/\S+@\S+\.\S+/.test(formData.email)) {
-      newErrors.email = t('auth.validation.email_invalid');
-    }
-
-    if (!formData.password) {
-      newErrors.password = t('auth.validation.required');
-    }
-
-    setErrors(newErrors);
-    return Object.keys(newErrors).length === 0;
-  };
+    remember: false,
+  });
+  const [showPassword, setShowPassword] = useState(false);
+  const [error, setError] = useState('');
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
-    
-    if (!validateForm()) return;
-
-    setIsLoading(true);
+    setError('');
 
     try {
-      await login(formData.email, formData.password);
-      
-      toast({
-        title: t('common.success'),
-        description: "Successfully logged in!",
-      });
-
-      // Redirect to the next page or dashboard
-      const nextUrl = searchParams.get('next') || '/dashboard';
-      navigate(nextUrl);
-    } catch (error) {
+      const success = await login({ email: formData.email, password: formData.password });
+      if (success) {
+        toast({
+          title: t('common.success'),
+          description: 'Successfully logged in!',
+        });
+        const nextUrl = searchParams.get('next') || '/dashboard';
+        navigate(nextUrl);
+      }
+    } catch (err: any) {
+      setError(err.message || 'Login failed');
       toast({
         title: t('common.error'),
-        description: "Invalid credentials. Please try again.",
-        variant: "destructive",
+        description: 'Invalid credentials. Please try again.',
+        variant: 'destructive',
       });
-    } finally {
-      setIsLoading(false);
     }
   };
 
   return (
     <div className="min-h-screen flex">
-      {/* Background Image Section */}
+      {/* Background */}
       <div className="hidden lg:flex lg:w-1/2 relative overflow-hidden">
-        <div 
+        <div
           className="absolute inset-0 bg-cover bg-center"
           style={{ backgroundImage: `url(${authBackground})` }}
         >
           <div className="absolute inset-0 bg-gradient-primary opacity-80"></div>
         </div>
-        
-        <div className="relative z-10 flex items-center justify-center p-12">
-          <div className="text-white text-center max-w-md">
-            <h1 className="text-4xl font-bold mb-4">
-              {t('brand.name')}
-            </h1>
-            <p className="text-xl opacity-90 leading-relaxed">
-              {t('brand.slogan')}
-            </p>
+        <div className="relative z-10 flex items-center justify-center p-12 text-white">
+          <div className="text-center max-w-md">
+            <h1 className="text-4xl font-bold mb-4">{t('brand.name')}</h1>
+            <p className="text-xl opacity-90">{t('brand.slogan')}</p>
           </div>
         </div>
       </div>
 
-      {/* Form Section */}
+      {/* Form */}
       <div className="w-full lg:w-1/2 flex items-center justify-center p-8 bg-background">
         <div className="w-full max-w-md space-y-6">
-          {/* Header */}
           <div className="flex justify-between items-center mb-8">
             <Link to="/" className="text-2xl font-bold gradient-text">
               {t('brand.name')}
@@ -122,63 +83,47 @@ const LoginPage: React.FC = () => {
             <LanguageToggle />
           </div>
 
-          {/* Login Card */}
           <Card className="glass-card animate-fade-in">
             <CardHeader className="text-center">
-              <CardTitle className="heading-md">
-                {t('auth.login.title')}
-              </CardTitle>
-              <CardDescription>
-                {t('auth.login.subtitle')}
-              </CardDescription>
+              <CardTitle className="heading-md">{t('auth.login.title')}</CardTitle>
+              <CardDescription>{t('auth.login.subtitle')}</CardDescription>
             </CardHeader>
-            
             <CardContent>
+              {error && (
+                <p className="text-sm text-destructive mb-2 flex items-center gap-1">
+                  <AlertCircle className="h-4 w-4" />
+                  {error}
+                </p>
+              )}
+
               <form onSubmit={handleSubmit} className="space-y-4">
-                {/* Email Field */}
-                <div className="space-y-2">
+                {/* Email */}
+                <div>
                   <Label htmlFor="email">{t('auth.login.email')}</Label>
                   <Input
                     id="email"
                     type="email"
                     value={formData.email}
                     onChange={(e) => setFormData({ ...formData, email: e.target.value })}
-                    className={errors.email ? 'border-destructive' : ''}
-                    placeholder="example@domain.com"
                   />
-                  {errors.email && (
-                    <p className="text-sm text-destructive flex items-center gap-1">
-                      <AlertCircle className="h-4 w-4" />
-                      {errors.email}
-                    </p>
-                  )}
                 </div>
 
-                {/* Password Field */}
-                <div className="space-y-2">
+                {/* Password */}
+                <div>
                   <Label htmlFor="password">{t('auth.login.password')}</Label>
                   <div className="relative">
-                      <Form {...form}>
-        <form
-          onSubmit={(e) => {
-            e.preventDefault();
-            form.handleSubmit(onSubmit)(e);
-          }}
-          className="space-y-4 w-full max-w-sm"
-        ></form>
                     <Input
                       id="password"
                       type={showPassword ? 'text' : 'password'}
                       value={formData.password}
                       onChange={(e) => setFormData({ ...formData, password: e.target.value })}
-                      className={errors.password ? 'border-destructive pr-10' : 'pr-10'}
-                      placeholder="••••••••"
+                      className="pr-10"
                     />
                     <Button
                       type="button"
                       variant="ghost"
                       size="sm"
-                      className="absolute right-0 top-0 h-full px-3 py-2 hover:bg-transparent"
+                      className="absolute right-0 top-0 h-full px-3"
                       onClick={() => setShowPassword(!showPassword)}
                     >
                       {showPassword ? (
@@ -188,39 +133,22 @@ const LoginPage: React.FC = () => {
                       )}
                     </Button>
                   </div>
-                  {errors.password && (
-                    <p className="text-sm text-destructive flex items-center gap-1">
-                      <AlertCircle className="h-4 w-4" />
-                      {errors.password}
-                    </p>
-                  )}
                 </div>
 
-                {/* Remember Me & Forgot Password */}
-                <div className="flex items-center justify-between">
-                  <div className="flex items-center space-x-2 rtl:space-x-reverse">
-                    <Checkbox
-                      id="remember"
-                      checked={formData.remember}
-                      onCheckedChange={(checked) => 
-                        setFormData({ ...formData, remember: checked as boolean })
-                      }
-                    />
-                    <Label htmlFor="remember" className="text-sm">
-                      {t('auth.login.remember')}
-                    </Label>
-                  </div>
-                  <Button variant="link" className="px-0 text-sm">
-                    {t('auth.login.forgot')}
-                  </Button>
+                {/* Remember */}
+                <div className="flex items-center space-x-2 rtl:space-x-reverse">
+                  <Checkbox
+                    id="remember"
+                    checked={formData.remember}
+                    onCheckedChange={(c) =>
+                      setFormData({ ...formData, remember: c as boolean })
+                    }
+                  />
+                  <Label htmlFor="remember">{t('auth.login.remember')}</Label>
                 </div>
 
-                {/* Submit Button */}
-                <Button
-                  type="submit"
-                  className="w-full glow-effect"
-                  disabled={isLoading}
-                >
+                {/* Submit */}
+                <Button type="submit" className="w-full glow-effect" disabled={isLoading}>
                   {isLoading ? (
                     <div className="flex items-center gap-2">
                       <div className="animate-spin rounded-full h-4 w-4 border-b-2 border-white"></div>
@@ -233,34 +161,14 @@ const LoginPage: React.FC = () => {
                     </>
                   )}
                 </Button>
-
-                {/* Register Link */}
-                <div className="text-center">
-                  <p className="text-sm text-muted-foreground">
-                    {t('auth.login.register_link').split('?')[0]}?{' '}
-                    <Link 
-                      to="/register" 
-                      className="text-primary hover:underline font-medium"
-                    >
-                      {t('auth.login.register_link').split('? ')[1]}
-                    </Link>
-                  </p>
-                </div>
               </form>
             </CardContent>
           </Card>
 
-          {/* Demo Credentials Alert */}
-          <Alert className="border-accent/20 bg-accent/10">
-            <AlertCircle className="h-4 w-4 text-accent" />
-            <AlertDescription className="text-accent-foreground">
-              <strong>Demo:</strong> Use any email and password to login (e.g., demo@avocat.com / password123)
-            </AlertDescription>
-          </Alert>
         </div>
       </div>
     </div>
   );
 };
 
-export default LoginPage;  
+export default LoginPage;
